requirements.txt
pandas==2.2.3
numpy==2.2.5
matplotlib==3.10.1
seaborn==0.13.2
geopandas==1.0.1
h3==4.2.2
scipy==1.15.2
statsmodels==0.14.4
inegipy==1.2.1
tqdm==4.67.1
scikit-learn==1.6.1
pyarrow==18.1.0
Jinja2==3.1.6
mapclassify==2.8.1
folium==0.19.5
python-dotenv==1.1.0
sie_banxico==0.0.1
ipykernel==6.29.5
PyYAML==6.0.2
boto3==1.38.7
awswrangler==3.11.0